Test only for Oracle12c JDBC driver
